Develop

+++ src/modules/club/entities/club.entity.ts
@@ -1,7 +1,6 @@
 import { Coach } from "src/modules/coach/entities/coach.entity";
 import { Player } from "src/modules/player/entities/player.entity";
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from "typeorm";
-import { UpdateClubDto } from "../dto/update-club.dto";
 
 @Entity()
 export class Club {
@@ -14,10 +13,10 @@ export class Club {
     })
     name: string;
 
-    @Column('numeric')
+    @Column('float')
     budget: number;
 
-    @Column('numeric')
+    @Column('float')
     remainingBudget?: number; 
 
     @OneToMany(() => Player, player => player.club)
